StreamReassembler ignore bytes that would extend beyond capacity.
Apparently this was fine for previous testcases, but is needed to pass
TCPConnection testcases that expect segments beyond the window size to
not be received.

The Reassembler lab description says that "Bytes that would exceed the
capacity are silently discarded". To me, this does not necessarily mean
out of window bytes should be discarded, but gotta do whatever to pass
edgecases.